Refactor truth code to be used by multiple scripts
