+++ data-truth/zoltar-truth-data.py
@@ -1,7 +1,6 @@
 import json
-from zoltpy import util
+from truth_utils._utils import get_epi_data, get_epi_data_TZ, get_available_timezeros, pre_process_truth, pre_process_epiweek, get_raw_truth_df, extract_raw_us_and_state_truth
 import pandas as pd
-import pymmwr as pm
 import datetime
 import warnings
 import requests
@@ -9,42 +8,8 @@ import io
 warnings.simplefilter(action='ignore')
 
 
-def get_epi_data(date):
-    format_str = '%m/%d/%y'  # The format
-    dt = datetime.datetime.strptime(date, format_str).date()
-    epi = pm.date_to_epiweek(dt)
-    return epi.year, epi.week, epi.day
-
-
-def get_epi_data_TZ(date):
-    format_str = '%Y-%m-%d'  # The format
-    dt = datetime.datetime.strptime(date, format_str).date()
-    epi = pm.date_to_epiweek(dt)
-    epi_week = epi.week
-    epi_day = epi.day
-    if epi_day >= 3:  # cut off is Tuesday
-        epi_week = epi_week + 1
-    return epi.year, epi_week, epi.day
-
-
-def get_available_timezeros(project_name):
-    conn = util.authenticate()
-    project = [project for project in conn.projects if project.name == project_name][0]
-    project_timezeros = project.timezeros
-    timezero = []
-    for timezero_array in project_timezeros:
-        timezero += [timezero_array.timezero_date]
-    return timezero
-
-
 def configure_JHU_data(df, target):
-    # convert matrix to repeating row format
-    df_truth = df.unstack()
-    df_truth = df_truth.reset_index()
-
-    # get epi data from date
-    df_truth['year'], df_truth['week'], df_truth['day'] = \
-        zip(*df_truth['level_0'].map(get_epi_data))
+    df_truth = pre_process_truth(df)
 
     # rename columns
     df_truth = df_truth.rename(columns={0: "value",
@@ -62,24 +27,8 @@ def configure_JHU_data(df, target):
     # add leading zeros to state code
     df_truth['location'] = df_truth['location'].apply(lambda x: '{0:0>2}'.format(x))
 
-    # Observed data on the seventh day
-    # or group by week for incident deaths
-    if target == 'Incident Deaths':
-        df_vis = df_truth.groupby(['week', 'location_long'], as_index=False).agg({'level_0': 'last',
-                                                                                  'value': 'sum',
-                                                                                  'year': 'last',
-                                                                                  'day': 'last',
-                                                                                  'location': 'last',
-                                                                                  'abbreviation': 'last'})
-    else:
-        df_vis = df_truth
-    df_vis['week'] = df_vis['week'] + 1  # shift epiweek on axis
-
-    # add leading zeros to epi week
-    df_vis['week'] = df_vis['week'].apply(lambda x: '{0:0>2}'.format(x))
-
-    # define epiweek
-    df_vis['epiweek'] = df_vis['year'].astype(str) + df_vis['week']
+    # Pre-process epiweek
+    df_vis = pre_process_epiweek(df_truth, target, for_zoltar = True)
 
     # rename columns
     df_truth_long = df_vis.rename(columns={"week": "epiweek",
@@ -139,24 +88,13 @@ def configure_JHU_data(df, target):
 
 
 url = "https://raw.githubusercontent.com/CSSEGISandData/COVID-19/master/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_US.csv"
-url_req = requests.get(url).content
-df = pd.read_csv(io.StringIO(url_req.decode('utf-8')))
+df = get_raw_truth_df(url)
 
 fips_codes = pd.read_csv('../data-locations/locations.csv')
 fips_codes = fips_codes[fips_codes.location != '11001']
 
-# aggregate by state and nationally
-state_agg = df.groupby(['Province_State']).sum()
-us_nat = df.groupby(['Country_Region']).sum()
-df_state_nat = state_agg.append(us_nat)
-
-# drop unnecessary columns
-cols = list(range(0, 6))
-df_truth = df_state_nat.drop(df_state_nat.columns[cols], axis=1)
-
-# calculate incidents from cumulative
-df_truth_cumulative = df_truth
-df_truth_incident = df_truth - df_truth.shift(periods=1, axis='columns')
+# Extraction of cumulative and incident truths for national and state level
+df_truth_cumulative, df_truth_incident = extract_raw_us_and_state_truth(df, for_zoltar = True)
 
 # re-format files
 df_cum_death = configure_JHU_data(df_truth_cumulative, "Cumulative Deaths")
